+++ lib/engine/game/g_18_eu/step/bankrupt.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require_relative '../../../step/bankrupt'
+
+module Engine
+  module Game
+    module G18EU
+      module Step
+        class Bankrupt < Engine::Step::Bankrupt
+          # TODO: Implement removing player from game
+        end
+      end
+    end
+  end
+end
