+++ lib/engine/game/g_18_eu/step/discard_train.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require_relative '../../../step/discard_train'
+
+module Engine
+  module Game
+    module G18EU
+      module Step
+        class DiscardTrain < Engine::Step::DiscardTrain
+          # TODO: Force discard Pullman first
+        end
+      end
+    end
+  end
+end
