+++ lib/engine/game/g_18_eu/step/merge.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+require_relative '../../../step/base'
+require_relative '../../../token'
+require_relative '../../../step/token_merger'
+
+module Engine
+  module Game
+    module G18EU
+      module Step
+        class Merge < Engine::Step::Base
+          include Engine::Step::TokenMerger
+
+          # TODO: Implement Mergers
+        end
+      end
+    end
+  end
+end
