--- lib/engine/step/g_18_eu/bankrupt.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../bankrupt'
-
-module Engine
-  module Step
-    module G18EU
-      class Bankrupt < Bankrupt
-        # TODO: Implement removing player from game
-      end
-    end
-  end
-end
