--- lib/engine/step/g_18_eu/discard_train.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../discard_train'
-
-module Engine
-  module Step
-    module G18EU
-      class DiscardTrain < DiscardTrain
-        # TODO: Force discard Pullman first
-      end
-    end
-  end
-end
